refactor: prepare for supporting different versions of compact divide & conquer
